Store

+++ services/townService/src/requestHandlers/PostCoveyTownRequestHandlers.ts
@@ -26,7 +26,7 @@ export interface PostGetRequest {
     postID : string
 }
 
-export interface PostGetIdInTownRequest {
+export interface PostGetAllInTownRequest {
     coveyTownID : string,
     sessionToken : string
 }
@@ -57,20 +57,17 @@ export interface CommentUpdateRequest {
     comment : Comment
 }
 
-const postTownController = new PostCoveyTownController("testTown", true);
-postTownController.addPlayer(new Player('test'));
+export async function postCreateHandler(_requestData : PostCreateRequest): Promise<ResponseEnvelope<Post | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
 
-export async function postCreateHandler(_requestData : PostCreateRequest): Promise<ResponseEnvelope<Post>> {
-    // const townsStore = PostCoveyTownStore.getInstance();
-    // const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
-
-    // if (!townController) {
-    //     return {
-    //         isOK: false,
-    //         response: undefined,
-    //         message: 'Unable to create post' 
-    //     }
-    // }
+    if (!postTownController) {
+        return {
+            isOK: false, 
+            response: {}, 
+            message: "Unable to create post",
+        };
+    }
 
     const post = _requestData.post;
     const result = await postTownController.createPost(post);
@@ -81,10 +78,41 @@ export async function postCreateHandler(_requestData : PostCreateRequest): Promi
     };
 }
 
-export async function postGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post>> {
-    
+export async function postGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post | Object>> {
+
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
     const postID = _requestData.postID;
+
+    if (!postTownController) {
+        return {
+            isOK: false,
+            response: {},
+            message: 'Unable to get post.'
+        };
+    }
+
     const result = await postTownController.getPost(postID);
+    return {
+        isOK: true,
+        response: result,
+        message: !result ? 'Invalid password. Please double check your town update password.' : undefined,
+    };
+
+}
+
+export async function postGetAllIDInTownHandler(_requestData : PostGetAllInTownRequest) : Promise<ResponseEnvelope<Post[] | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController) {
+        return {
+            isOK: false,
+            response: {},
+            message: 'Unable to get all Posts in Town.'
+        };
+    }
+    const result : Post[] = await postTownController.getAllPostInTown();
 
     return {
         isOK: true,
@@ -93,17 +121,17 @@ export async function postGetHandler(_requestData : PostGetRequest) : Promise<Re
     };
 }
 
-export async function postGetAllIDInTownHandler(_requestData : PostGetIdInTownRequest) : Promise<ResponseEnvelope<string[]>> {
-    const result : string[] = await postTownController.getAllPostInTown();
+export async function postGetCommentTreeHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<CommentTree[] | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
 
-    return {
-        isOK: true,
-        response: result,
-        message: !result ? 'Invalid password. Please double check your town update password.' : undefined,
-    };
-}
-
-export async function postGetCommentTreeHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<CommentTree[]>> {
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to delete post with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
+        };
+    }
     const result: CommentTree[] = await postTownController.getCommentTree(_requestData.postID);
 
     return {
@@ -113,11 +141,15 @@ export async function postGetCommentTreeHandler(_requestData : PostGetRequest) :
     }
 }
 
-export async function postDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post>> {
-    if (!postTownController?.getSessionByToken(_requestData.sessionToken)){
+export async function postDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<Post | Object>> {
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
         return {
           isOK: false, 
-          response: undefined, 
+          response: {}, 
           message: `Unable to delete post with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
         };
     }
@@ -132,9 +164,20 @@ export async function postDeleteHandler(_requestData : PostGetRequest) : Promise
     };
 }
 
-export async function postUpdateHandler(_requestData : PostUpdateRequest) : Promise<ResponseEnvelope<Post>> {
+export async function postUpdateHandler(_requestData : PostUpdateRequest) : Promise<ResponseEnvelope<Post | Object>> {
     const postID = _requestData.postID;
     const post = _requestData.post;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to update post with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
+        };
+    }
 
     const result = await postTownController.updatePost(postID, post, _requestData.sessionToken);
 
@@ -145,9 +188,20 @@ export async function postUpdateHandler(_requestData : PostUpdateRequest) : Prom
     };
 }
 
-export async function commentCreateHandler(_requestData : CommentCreateRequest): Promise<ResponseEnvelope<Comment>> {
+export async function commentCreateHandler(_requestData : CommentCreateRequest): Promise<ResponseEnvelope<Comment | Object>> {
     const comment = _requestData.comment;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to create comment`,
+        };
+    }
     const result = await postTownController.createComment(comment);
+    
     return {
         isOK: true,
         response: result,
@@ -155,9 +209,19 @@ export async function commentCreateHandler(_requestData : CommentCreateRequest):
     };
 }
 
-export async function commentGetHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentGetHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     
     const commentID = _requestData.commentID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to get comment`,
+        };
+    }
     const result = await postTownController.getComment(commentID);
 
     return {
@@ -167,8 +231,19 @@ export async function commentGetHandler(_requestData : CommentGetRequest) : Prom
     };
 }
 
-export async function commentDeleteHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentDeleteHandler(_requestData : CommentGetRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     const commentID = _requestData.commentID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to delete comment with comment ID ${_requestData.commentID} in town ${_requestData.coveyTownID}`,
+        };
+    }
     const result = await postTownController.deleteComment(commentID, _requestData.sessionToken);
 
     return {
@@ -178,9 +253,20 @@ export async function commentDeleteHandler(_requestData : CommentGetRequest) : P
     };
 }
 
-export async function commentUpdateHandler(_requestData : CommentUpdateRequest) : Promise<ResponseEnvelope<Comment>> {
+export async function commentUpdateHandler(_requestData : CommentUpdateRequest) : Promise<ResponseEnvelope<Comment | Object>> {
     const commentID = _requestData.commentID;
     const comment = _requestData.comment;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to update comment with comment ID ${_requestData.commentID} in town ${_requestData.coveyTownID}`,
+        };
+    }
 
     const result = await postTownController.updateComment(commentID, comment, _requestData.sessionToken);
 
@@ -193,6 +279,16 @@ export async function commentUpdateHandler(_requestData : CommentUpdateRequest) 
 
 export async function fileGetHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<any>> {
     const postID = _requestData.postID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+
+    if (!postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to get file`,
+        };
+    }
     const result = await postTownController.getFile(postID);
 
     return {
@@ -204,6 +300,18 @@ export async function fileGetHandler(_requestData : PostGetRequest) : Promise<Re
 
 export async function fileDeleteHandler(_requestData : PostGetRequest) : Promise<ResponseEnvelope<any>> {
     const postID = _requestData.postID;
+    const townsStore = CoveyTownsStore.getInstance();
+    const postTownController = townsStore.getPostControllerForTown(_requestData.coveyTownID);
+    const townController = townsStore.getControllerForTown(_requestData.coveyTownID);
+
+    if (!townController?.getSessionByToken(_requestData.sessionToken) || !postTownController){
+        return {
+          isOK: false, 
+          response: {}, 
+          message: `Unable to delete file with post ID ${_requestData.postID} in town ${_requestData.coveyTownID}`,
+        };
+    }
+
     const result = await postTownController.deleteFile(postID, _requestData.sessionToken);
 
     return {
